pyproject.toml
[project]
name = "kaspa-trading-bot"
version = "0.1.0"
description = "Kaspa trading bot with ccxt integrations and Kaspa RPC support."
authors = [{ name = "Kaspa Ops" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.5",
    "aioredis>=2.0.1",
    "asyncpg>=0.29.0",
    "ccxt>=4.3.35",
    "httpx>=0.27.0",
    "prometheus-client>=0.20.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "redis>=5.0.8",
    "structlog>=24.1.0",
    "tenacity>=9.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.3.0",
    "ruff>=0.6.8",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
target-version = "py311"

